% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\name{aes}
\alias{aes}
\title{Modified Aesthetic Mappings}
\usage{
aes(x, y, z, ...)
}
\arguments{
\item{x}{x value}

\item{y}{y value}

\item{z}{z value}

\item{...}{other arguments as per \code{\link[ggplot2]{aes}}}
}
\description{
Modified Aesthetic Mappings
}
\details{
An extension to the base aes functin from ggplot2, this is modified to handle a default z mapping 
for application in ternary phase diagrams. Does not alter the standard behaviour.
}
\seealso{
Parent \code{\link[ggplot2]{aes}} function.
}
