% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strip-unapproved.R
\name{approved_geometries}
\alias{approved_geometries}
\title{Approved Geometries}
\description{
\Sexpr[results=rd,stage=build]{ggtern:::.rd_approved(TRUE)}
\Sexpr[results=rd,stage=build]{ggtern:::.rd_approved(FALSE)}
}
\details{
The balance of the available geometries to ggplot2 are work in progress with regards to the \code{ggtern} package.

Attempting to apply non-approved geometries (ie geometries not in the above list), will result in them being stripped from
the final plot.
}

