% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications-gridExtra.R
\name{arrangeGrob}
\alias{arrangeGrob}
\alias{grid.arrange}
\title{Arrange multiple grobs on a page (ggtern version)}
\usage{
arrangeGrob(
  ...,
  grobs = list(...),
  layout_matrix,
  vp = NULL,
  name = "arrange",
  as.table = TRUE,
  respect = FALSE,
  clip = "off",
  nrow = NULL,
  ncol = NULL,
  widths = NULL,
  heights = NULL,
  top = NULL,
  bottom = NULL,
  left = NULL,
  right = NULL,
  padding = unit(0.5, "line")
)

grid.arrange(..., newpage = TRUE)
}
\arguments{
\item{...}{grobs, gtables, ggplot or trellis objects}

\item{grobs}{list of grobs}

\item{layout_matrix}{optional layout}

\item{vp}{viewport}

\item{name}{argument of gtable}

\item{as.table}{logical: bottom-left to top-right (TRUE) or top-left to bottom-right (FALSE)}

\item{respect}{argument of gtable}

\item{clip}{argument of gtable}

\item{nrow}{argument of gtable}

\item{ncol}{argument of gtable}

\item{widths}{argument of gtable}

\item{heights}{argument of gtable}

\item{top}{optional string, or grob}

\item{bottom}{optional string, or grob}

\item{left}{optional string, or grob}

\item{right}{optional string, or grob}

\item{padding}{unit of length one, margin around annotations}

\item{newpage}{open a new page}
}
\description{
A very slight modification to the original function, removing the explicit direction to use the ggplotGrob function
from the ggplot2 namespace
}
\author{
Nicholas Hamilton
}
