\name{arrow}
\alias{arrow}
\title{\code{arrow} is an exported instance of the imported grid arrow function.
For further information, refer to: \link[grid]{arrow}.}
\description{
  \code{arrow} is an exported instance of the imported grid
  arrow function. For further information, refer to:
  \link[grid]{arrow}.
}

