% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{breaks_tern}
\alias{breaks_tern}
\alias{getBreaks}
\title{Generate Axis Breaks}
\usage{
breaks_tern(limits = c(0, 1), isMajor = TRUE, n = 5)
}
\arguments{
\item{limits}{the scale limits}

\item{isMajor}{major or minor grids}

\item{n}{number of breaks}
}
\description{
Calculates the Breaks for Major or Minor Gridlines based on the input limits.
}
\examples{
 breaks_tern()
 breaks_tern(limits = c(0,.5),FALSE,10)
}
