% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clip-polygons.R
\name{clipPolygons}
\alias{clipPolygons}
\alias{polyclip}
\title{Clip Polygons}
\usage{
clipPolygons(df, coord, plyon = c("level", "piece", "group"),
  op = "intersection")
}
\arguments{
\item{df}{a data frame}

\item{coord}{a ternary coordinate system}

\item{plyon}{items in the data frame to pass to ddply argument}

\item{op}{operation method to clip, intersection, union, minus or xor}
}
\description{
Using the using the PolyClip Package, This clips input polygons for
use in the density and contour geometries.
}
\keyword{clipping}
\keyword{polygon}

