% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{data}
\alias{data}
\alias{ggtern-data}
\title{Data Sets}
\description{
\code{ggtern} ships with a number of datasets, including:
\enumerate{
  \item \code{\link[=data_Feldspar]{Elkin and Groves Feldspar Data}}
  \item \code{\link[=data_USDA]{USDA Textural Classification Data}}
}
}

