% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\name{data_ggtern_package}
\alias{data_ggtern_package}
\title{Data Sets}
\description{
\code{ggtern} ships with a number of datasets, including:
\enumerate{
  \item \code{\link[=data_Feldspar]{Elkin and Groves Feldspar Data}}
  \item \code{\link[=data_USDA]{USDA Textural Classification Data}}
  \item \code{\link[=data_Fragments]{Grantham and Valbel Rock Fragment Data}}
}
}
\author{
Nicholas Hamilton
}

