% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{data_sets_Feldspar}
\alias{Feldspar}
\alias{data_sets_Feldspar}
\title{Elkin and Groves Feldspar Data}
\format{One (1) row per Feldspar composition}
\usage{
data(Feldspar)
}
\description{
This dataset represents the experiments conducted by Elkins and Groves
}
\examples{
 data(Feldspar)
 summary(Feldspar)
 ggtern(data=Feldspar,aes(x=An,y=Ab,z=Or)) + geom_point()
}
\author{
Nicholas Hamilton
}
\references{
Elkins, L. T. & Grove, T. L. 
Ternary Feldspar Experiments and Thermodynamic Models 
American Mineralogist, Mineral Soc America, 1990, 75, 544-559
}
\seealso{
\link[=data]{ggtern datasets}
}

