% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend-draw-tern.R
\name{draw_key_tern}
\alias{draw_key_tern}
\alias{draw_key_crosshair_tern}
\alias{draw_key_Tmark}
\alias{draw_key_Lmark}
\alias{draw_key_Rmark}
\alias{draw_key_Tline}
\alias{draw_key_Lline}
\alias{draw_key_Rline}
\alias{draw_key_Tiso}
\alias{draw_key_Liso}
\alias{draw_key_Riso}
\alias{draw_key_point_swap}
\title{Key drawing functions}
\usage{
draw_key_crosshair_tern(data, params, size)

draw_key_Tmark(data, params, size)

draw_key_Lmark(data, params, size)

draw_key_Rmark(data, params, size)

draw_key_Tline(data, params, size)

draw_key_Lline(data, params, size)

draw_key_Rline(data, params, size)

draw_key_Tiso(data, params, size)

draw_key_Liso(data, params, size)

draw_key_Riso(data, params, size)

draw_key_point_swap(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\value{
A grid grob.
}
\description{
Each Geom has an associated function that draws the key when the geom needs
to be displayed in a legend. These are the options built into ggplot2.
}
\author{
Nicholas Hamilton
}
