% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{expandRange}
\alias{expandRange}
\title{Expand a Range of Values}
\usage{
expandRange(x, m = 1)
}
\arguments{
\item{x}{range of values, required to be numeric and of length 2}

\item{m}{multiplyer}
}
\description{
Expands a range of values about its midpoint, by an amount equal to the multiplyer
}

