% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{expandTern}
\alias{expandTern}
\title{Determine Ternary Expansion}
\usage{
expandTern(coord, by = getOption("tern.expand"))
}
\arguments{
\item{coord}{ternary coordinates}

\item{by}{fraction to expand by}
}
\value{
numeric scalar, of the amount to expand ternary grid
}
\description{
Determines the Expansion Buffer on the Ternary Sufrace,
Similar to the 'expand' term in ggplot2 on a rectangular grid
}

