\name{geom_segment_tern}
\alias{geom_segment_tern}
\title{Geom Segment for Ternary System}
\usage{
  geom_segment_tern(mapping = NULL, data = NULL,
    stat = "identity", position = "identity", arrow = NULL,
    lineend = "butt", na.rm = FALSE, ...)
}
\description{
  This is the ternary equivalent for the geom_segment
  proto, requiring \code{zend} and \code{z} additional
  mandatory aesthetic mappings.
}
\seealso{
  \code{\link{geom_segment}}
}

