\docType{data}
\name{GeomSegment}
\alias{GeomSegment}
\alias{ggtern-internal}
\title{Internal Function}
\format{proto object 
 $ draw        :function (., data, scales, coordinates, arrow = NULL, lineend = "butt", na.rm = FALSE, ...)   
 $ default_stat:function (.)   
 $ objname     : chr "segment" 
 $ default_aes :function (.)   
 $ guide_geom  :function (.)   
 $ required_aes: chr [1:2] "x" "y" 
 parent: proto object 
 .. parent: proto object}
\usage{
  GeomSegment
}
\description{
  Internal Function
}
\keyword{datasets}

