\docType{data}
\name{GeomSmoothMod}
\alias{GeomSmoothMod}
\title{Modified Geom Smooth Proto}
\format{proto object 
 $ draw        :function (., data, scales, coordinates, limitarea = F, ...)   
 $ default_stat:function (.)   
 $ objname     : chr "smooth" 
 $ default_aes :function (.)   
 $ guide_geom  :function (.)   
 $ required_aes: chr [1:2] "x" "y" 
 $ draw_legend :function (., data, params, ...)   
 parent: proto object 
 .. parent: proto object}
\usage{
  GeomSmoothMod
}
\description{
  Modified Geom Smooth Proto
}
\keyword{datasets}

