% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{getBreaks}
\alias{getBreaks}
\title{Get Number of Breaks}
\usage{
getBreaks(limits, isMajor, nMajor = 5, nMinor = 2 * nMajor)
}
\arguments{
\item{limits}{the scale limits}

\item{isMajor}{major or minor grids}

\item{nMajor}{number of major breaks}

\item{nMinor}{number of minor breaks}
}
\description{
Calculates the Breaks for Major or Minor Gridlines
based on the input limits.
}

