\name{ggtern}
\alias{ggtern}
\title{Create ggtern object}
\usage{
  ggtern(...)
}
\arguments{
  \item{...}{same arguments as passed through to ggplot}
}
\description{
  \code{ggtern} is a function, analogous to \code{ggplot},
  which creates an object of class \code{ggtern} (which
  inherits \code{ggplot}), and creation of a ternary plot
  is predicated on using this constructor.
}
\seealso{
  \code{\link{ggplot}}
}

