\name{ggtern-package}
\alias{ggtern-package}
\alias{ggtern}
\docType{package}
\title{
An extension to ggplot2 for ternary diagrams
}
\description{
Ternary diagrams are used frequently in materials science to graph compositional features for mixtures of three different elements.
}
\details{
\tabular{ll}{
Package: \tab ggtern\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-15\cr
License: \tab GPL-2\cr
}
Overview to come
}
\author{
Written and Maintained by Nicholas Hamilton <n.hamilton@student.unsw.edu.au>
}
\references{
Literature to come
}
\keyword{ package }
\seealso{

  Get started: \code{\link{ggtern_constructor}}, and \code{\link{ggplot2}}.
}
