% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-defaults.R
\name{ggtern_themes}
\alias{ggtern_themes}
\alias{theme_ggtern}
\alias{theme_gray}
\alias{theme_bw}
\alias{theme_linedraw}
\alias{theme_light}
\alias{theme_minimal}
\alias{theme_classic}
\alias{theme_dark}
\alias{theme_void}
\alias{theme_darker}
\alias{theme_custom}
\alias{theme_rgbw}
\alias{theme_rgbg}
\alias{theme_matrix}
\alias{theme_tropical}
\alias{theme_bluedark}
\alias{theme_bluelight}
\alias{theme_bvbw}
\alias{theme_bvbg}
\title{ggtern themes}
\usage{
theme_ggtern(base_size = 11, base_family = "")

theme_gray(base_size = 11, base_family = "")

theme_bw(base_size = 12, base_family = "")

theme_linedraw(base_size = 12, base_family = "")

theme_light(base_size = 12, base_family = "")

theme_minimal(base_size = 12, base_family = "")

theme_classic(base_size = 12, base_family = "")

theme_dark(base_size = 12, base_family = "")

theme_void(base_size = 12, base_family = "")

theme_darker(base_size = 12, base_family = "")

theme_custom(
  base_size = 12,
  base_family = "",
  tern.plot.background = NULL,
  tern.panel.background = NULL,
  col.T = "black",
  col.L = "black",
  col.R = "black",
  col.grid.minor = "white"
)

theme_rgbw(base_size = 12, base_family = "")

theme_rgbg(base_size = 12, base_family = "")

theme_matrix(base_size = 12, base_family = "")

theme_tropical(base_size = 12, base_family = "")

theme_bluedark(base_size = 12, base_family = "")

theme_bluelight(base_size = 12, base_family = "")

theme_bvbw(base_size = 12, base_family = "")

theme_bvbg(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{tern.plot.background}{colour of background colour to plot area}

\item{tern.panel.background}{colour of panel background of plot area}

\item{col.T}{colour of top axis, ticks labels and major gridlines}

\item{col.L}{colour of left axis, ticks, labels and major gridlines}

\item{col.R}{colour of right axis, ticks, labels and major gridlines}

\item{col.grid.minor}{the colour of the minor grid
\code{theme_custom} is a convenience function to allow the user to control the basic theme colours very easily.}
}
\description{
Themes set the general aspect of the plot such as the colour of the
background, gridlines, the size and colour of fonts.
}
\details{
\describe{

\item{\code{theme_gray}}{
The signature ggplot2 theme with a grey background and white gridlines,
designed to put the data forward yet make comparisons easy.
}

\item{\code{theme_bw}}{
The classic dark-on-light ggplot2 theme. May work better for presentations
displayed with a projector.
}

\item{\code{theme_linedraw}}{
A theme with only black lines of various widths on white backgrounds,
reminiscent of a line drawings. Serves a purpose similar to \code{theme_bw}.
Note that this theme has some very thin lines (<< 1 pt) which some journals
may refuse.}

\item{\code{theme_light}}{
A theme similar to \code{theme_linedraw} but with light grey lines and axes,
to direct more attention towards the data.
}

\item{\code{theme_dark}}{
The dark cousin of \code{theme_light}, with similar line sizes but a dark background. 
Useful to make thin coloured lines pop out.
}

\item{\code{theme_darker}}{
A darker cousing to \code{theme_dark}, with a dark panel background.
}

\item{\code{theme_minimal}}{
A minimalistic theme with no background annotations.
}

\item{\code{theme_classic}}{
A classic-looking theme, with x and y axis lines and no gridlines.
}

\item{\code{theme_rgbw}}{
 A theme with white background, red, green and blue axes and gridlines
}

\item{\code{theme_rgbg}}{
A theme with grey background, red, green and blue axes and gridlines
}

\item{\code{theme_void}}{ 
A completely empty theme.
}

\item{\code{theme_custom}}{
Theme with custom basic colours
}

\item{\code{theme_matrix}}{
Theme with very dark background and bright green features
}

\item{\code{theme_tropical}}{
Theme with tropical colours
}

\item{\code{theme_bluelight}}{
A blue theme with light background and dark features
}

\item{\code{theme_bluedark}}{
A blue theme with dark background and light features
}

\item{\code{theme_bvbw}}{
A black/vermillion/blue theme with white background, for colorblind sensitive readers, see references.
}

\item{\code{theme_bvbg}}{
A black/vermillion/blue theme with grey background, for colorblind sensitive readers, see references.
}

}
}
\examples{
\donttest{
#Create a list of the theme suffixes
themesOrg = c('gray','bw','linedraw','light',
              'dark','minimal','classic','void')
themesNew = c('custom','darker','rgbw','rgbg','tropical',
              'matrix','bluelight','bluedark','bvbw','bvbg')

#Iterate over all the suffixes, creating a list of plots
plotThemes = function(themes){
   grobs = lapply(themes,function(x){
     thmName = sprintf("theme_\%s",x)
     thm = do.call(thmName,args=list(base_size=9))
     df  = data.frame(label=thmName)
     ggtern(df) + facet_wrap(~label) + thm
   })
   grobs
}

#Arrange the Original Themes
grid.arrange(grobs=plotThemes(themesOrg),top = "Collection of Themes (Original)")

#Arrange the New Themes
grid.arrange(grobs=plotThemes(themesNew),top = "Collection of Themes (New Themes)")
}

}
\references{
Okabe, Masataka, and Kei Ito. "How to make figures and presentations that are friendly to color blind people." 
University of Tokyo (2002). http://jfly.iam.u-tokyo.ac.jp/color/
}
\author{
Nicholas Hamilton
}
