\name{iflasttern}
\alias{iflasttern}
\title{Switch between two values for the ternary system}
\usage{
iflasttern(yes = stop("yes value required"), no = stop("no value required"))
}
\arguments{
  \item{yes}{value to return when the current system is
  ternary}

  \item{no}{value to return when the current system is not}
}
\description{
Returns yes if the current system is ternary, and no if it
is not
}
\keyword{internal}

