% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{label_formatter}
\alias{label_formatter}
\title{\code{label_formatter} is a function that formats / parses labels for use in the grid.}
\usage{
label_formatter(label, ...)
}
\arguments{
\item{label}{character label}

\item{...}{additional arguments}
}
\description{
\code{label_formatter} is a function that formats / parses labels for use in the grid.
}
