% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{labels_tern}
\alias{labels_tern}
\alias{getLabels}
\title{Generate Axis Labels}
\usage{
labels_tern(
  limits = c(0, 1),
  breaks = breaks_tern(limits),
  format = "\%g",
  factor = 100
)
}
\arguments{
\item{limits}{the scale limits}

\item{breaks}{numeric denoting the breaks to produce corresponding labels}

\item{format}{the formatting string to be passed through to the \code{\link{sprintf}} function}

\item{factor}{the multiplicative factor}
}
\description{
Calculates the Labels for Major or Minor Gridlines based on the input limits.
}
\examples{
labels_tern()
labels_tern(limits = c(0,.5))
}
\author{
Nicholas Hamilton
}
