\name{get_last_coord}
\alias{get_last_coord}
\alias{set_last_coord}
\title{Last Coordinates}
\usage{
  get_last_coord()

  set_last_coord(new)
}
\arguments{
  \item{new}{the coordinate system to set}
}
\value{
  coordinate system
}
\description{
  Internal functions to get and set the last coordinate
  system. An internal command for ggtern.

  \code{set_last_coord} set the last coordinate system.
}
\details{
  \code{get_last_coord} is a function that returns the last
  coordinate system used.
}

