\name{isomLR}
\alias{isomLR}
\alias{isomLRinv}
\alias{log-ratio}
\alias{logratio}
\title{Log-ratio Transformations}
\usage{
  isomLR(x)

  isomLRinv(x.ilr)
}
\arguments{
  \item{x}{2D object to transform}

  \item{x.ilr}{2D object to transform}
}
\value{
  transformed data

  transformed data
}
\description{
  Log-ratio transformation, algorithm based on that
  provided in \code{\link[robCompositions]{isomLR}}

  Inverse isometric log-ratio transformation, algorithm
  based on that provided in
  \code{\link[robCompositions]{isomLRinv}}
}
\references{
  provided in \code{\link[robCompositions]{isomLR}}
}

