\name{notransform}
\alias{notransform}
\title{\code{notransform} is an internal function that permits the input argument (evaluate) to be executed with temporary disabling of
ternary transformations, existing state is restored}
\usage{
notransform(evaluate)
}
\arguments{
  \item{evaluate}{code to execute in a protected block from
  transformation}
}
\description{
\code{notransform} is an internal function that permits the
input argument (evaluate) to be executed with temporary
disabling of ternary transformations, existing state is
restored
}
\keyword{internal}

