% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-construction.R
\name{+.gg}
\alias{+.gg}
\alias{\%+\%}
\title{Plot Construction}
\usage{
\method{+}{gg}(e1, e2)

e1 \%+\% e2
}
\arguments{
\item{e1}{first object}

\item{e2}{second object}
}
\description{
\code{"+.gg"} is a local copy of the ggplot2 add function, no change other than exporting from the ggtern namespace
}
\author{
Nicholas Hamilton
}
