% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-nudge-tern.R
\docType{data}
\name{position_nudge_tern}
\alias{position_nudge_tern}
\alias{PositionNudgeTern}
\title{Nudge Ternary Points.}
\usage{
position_nudge_tern(x = 0, y = 0, z = 0)
}
\arguments{
\item{x, y, z}{Amount of compositions to nudge}
}
\description{
This is useful if you want to nudge labels a little ways from their points, input data will normalised to sum to unity before applying the particular
nudge, so the nudge variables should be as a fraction ie (0,1)
}
\seealso{
Other position adjustments: 
\code{\link{position_jitter_tern}()}
}
\author{
Nicholas Hamilton
}
\concept{position adjustments}
\keyword{datasets}
