% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-interpolate-methods.R
\name{predictdf2d}
\alias{predictdf2d}
\title{Prediction data frame}
\usage{
predictdf2d(model, xseq, yseq)
}
\arguments{
\item{model}{the model to predict}

\item{xseq, yseq}{the x and y values}
}
\description{
Get predictions with standard errors into data frame
}
