\name{print.ggtern}
\alias{plot.ggtern}
\alias{print.ggtern}
\title{Draw plot on current graphics device.}
\usage{
  \method{print}{ggplot} (x, newpage = is.null(vp),
    vp = NULL, ..., verbose = T)

  \method{plot}{ggplot} (x, newpage = is.null(vp),
    vp = NULL, ..., verbose = T)
}
\arguments{
  \item{x}{plot to display}

  \item{newpage}{draw new (empty) page first?}

  \item{vp}{viewport to draw plot in}

  \item{...}{other arguments not used by this method}
}
\description{
  Draw plot on current graphics device.
}
\keyword{hplot}

