\name{theme_nogrid_minor}
\alias{tern_hidegrid}
\alias{tern_nogrid}
\alias{tern_nogrid_major}
\alias{tern_nogrid_minor}
\alias{theme_hidegrid}
\alias{theme_nogrid}
\alias{theme_nogrid_major}
\alias{theme_nogrid_minor}
\alias{theme_tern_nogrid}
\alias{theme_tern_nogrid_major}
\alias{theme_tern_nogrid_minor}
\title{Show or Hide Grid}
\usage{
  theme_nogrid_minor()

  theme_nogrid_major()

  theme_nogrid()
}
\description{
  \code{theme_nogrid_minor} ternary theme, no minor grids.

  \code{theme_nogrid_major} ternary theme, no major grids.

  \code{theme_nogrid} ternary theme, no major AND no minor
  grids.
}

