% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stat-density-tern.R
\name{stat_density_tern}
\alias{StatDensityTern}
\alias{stat_density_tern}
\title{2d density estimation - (ggtern version)}
\usage{
stat_density_tern(mapping = NULL, data = NULL, stat = "DensityTern",
  position = "identity", na.rm = FALSE, contour = TRUE,
  n = getOption("tern.mesh.size"), buffer = getOption("tern.mesh.buffer"),
  ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link{aes}} or \code{\link{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{stat}{The statistical transformation to use on the data for this
layer.}

\item{position}{The position adjustment to use for overlapping points
on this layer}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{contour}{display the contour or show the mesh}

\item{n}{Number of grid points in each direction of the mesh. Can be scalar or a length-2 integer vector.}

\item{buffer}{factor to buffer the mesh, to prevent ugly truncation of contours, 1.0 means no buffering}

\item{...}{other arguments passed on to \code{\link{layer}}. This can
include aesthetics whose values you want to set, not map. See
\code{\link{layer}} for more details.}
}
\description{
Patched version of the 2d density estimation.
}
\seealso{
\code{\link[ggtern]{geom_density_tern}}
}

