% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stat-interpolate-tern.R
\name{stat_interpolate_tern}
\alias{StatInterpolateTern}
\alias{stat_interpolate_tern}
\title{Ternary Interpolation}
\usage{
stat_interpolate_tern(mapping = NULL, data = NULL,
  geom = "InterpolateTern", position = "identity", na.rm = FALSE,
  buffer = getOption("tern.mesh.buffer"), ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link{aes}} or \code{\link{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlappling points
on this layer}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{buffer}{factor to buffer the mesh, to prevent ugly truncation of contours, 1.0 means no buffering}

\item{...}{other arguments passed on to the \code{method}, such as  \code{\link{lm}} (default) or \code{\link{loess}}}
}
\value{
A data frame with additional column:
 \item{level}{height of contour}
}
\description{
This is the heavily requested statistic for interpolating between ternary values, results being
rendered using contours on a ternary mesh
}
\details{
By default, the interpolation is done using multivariate linear regression using the
following \code{formula=value~poly(x,y,z,degree=2)},
however this can be changed to anything that suits the user by including \code{method} and/or \code{formula} arguments
which get passed through to the model fitting and prediction components of the calculation routine.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{ggtern:::rd_aesthetics("stat", "InterpolateTern")}
}
\seealso{
\code{\link{geom_interpolate_tern}}, \code{\link{lm}}, \code{\link{loess}}
}

