\docType{data}
\name{StatDensity2d}
\alias{StatDensity2d}
\alias{ggtern-internal}
\title{Internal Function}
\format{proto object 
 $ objname     : chr "density2d" 
 $ default_aes :function (.)   
 $ default_geom:function (.)   
 $ calculate   :function (., data, scales, na.rm = FALSE, contour = TRUE, n = 100, geometry = "density2d", ...)   
 $ required_aes: chr [1:2] "x" "y" 
 parent: proto object 
 .. parent: proto object}
\usage{
  StatDensity2d
}
\description{
  Internal Function
}
\keyword{datasets}

