% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{suppressColours}
\alias{suppressColours}
\title{Suppress Colours}
\usage{
suppressColours(data, coord, toColor = "transparent", remove = FALSE)
}
\arguments{
\item{data}{ggplot dataframe}

\item{coord}{ternary coordinates}

\item{toColor}{the colour to suppress to}

\item{remove}{should suppressed points be removed entirely}
}
\description{
Function to Suppress Colours
}

