% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tern-stop.R
\name{tern_stop}
\alias{tern_stop}
\title{Stop Procedure}
\usage{
tern_stop(src = "target")
}
\arguments{
\item{src}{character name of current procedure}
}
\description{
Internal Function, checks if the most recent coordinate system is ternary, and, if not,
stops the current procedure, with a common message format
}
\keyword{internal}

