% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{ternLimitsForCoord}
\alias{ternLimitsForCoord}
\title{Ternary Limits}
\usage{
ternLimitsForCoord(coord)
}
\arguments{
\item{coord}{ternary coordinates}
}
\description{
Determine the Ternary Limits for Coordinate System
}

