\name{theme_nocart}
\alias{theme_nocart}
\title{Ternary Theme Elements}
\usage{
  theme_nocart()
}
\description{
  \code{theme_nocart} is a functino that returns empty
  theme elemens for the standard cartesian memebers,
  including panel.background, panel.border,
  panel.grid.major, panel.grid.minor, axis.ticks,
  axis.text.x, axis.text.y, axis.title.x, axis.title.y are
  all set to blank
}

