% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-bordersontop.R
\name{theme_bordersontop}
\alias{theme_bordersontop}
\alias{theme_bordersonbottom}
\title{Render Borders on Top}
\usage{
theme_bordersontop()

theme_bordersonbottom()
}
\description{
Convenience functions to render the axis border lines on top (or bottom) of the other layers. 
By default the borders are rendered in the background (bottom)
}
\author{
Nicholas Hamilton
}
