\name{theme_bw}
\alias{theme_bw}
\alias{theme_tern_bw}
\title{A theme with white background and black gridlines (ggtern version)}
\usage{
theme_bw(base_size = 12, base_family = "")
}
\arguments{
  \item{base_size}{base font size}

  \item{base_family}{base font family}
}
\description{
\code{theme_bw} is a theme with white background and black
gridlines
}

