% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme-defaults.R
\name{theme_classic}
\alias{theme_classic}
\alias{theme_tern_classic}
\title{A classic-looking theme, with x and y axis lines and no gridlines (ggtern version).}
\usage{
theme_classic(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
A classic-looking theme, with x and y axis lines and no gridlines (ggtern version).
}

