% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme-convenience.R
\name{theme_complete}
\alias{theme_complete}
\title{Complete Themes}
\description{
\code{ggtern} ships with a number of complete themes:
\itemize{
  \item Black and White Theme: \code{\link[=theme_tern_bw]{theme_bw(...)}}
  \item Minimal Theme: \code{\link[=theme_tern_minimal]{theme_minimal(...)}}
  \item Classic Theme: \code{\link[=theme_tern_classic]{theme_classic(...)}}
  \item Gray and White Theme: \code{\link[=theme_tern_gray]{theme_gray(...)}}
  \item Red, Green, Blue and White Theme: \code{\link[=theme_tern_rgbw]{theme_rgbw(...)}}
  \item Red, Green, Blue and Gray Theme: \code{\link[=theme_tern_rgbg]{theme_rgbg(...)}}
}
}

