% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme-defaults.R
\name{theme_gray}
\alias{theme_gray}
\alias{theme_grey}
\alias{theme_tern_gray}
\title{A theme with grey background and white gridlines (ggtern version)}
\usage{
theme_gray(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
\code{theme_gray} is a theme with grey background and white gridlines.
}

