% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-gridsontop.R
\name{theme_gridsontop}
\alias{theme_gridsontop}
\alias{theme_gridsonbottom}
\title{Render Grids on Top}
\usage{
theme_gridsontop()

theme_gridsonbottom()
}
\description{
Convenience function to render the major and minor grids on top (or bottom) of the other layers.
By default the grids are rendered in the background (bottom)
}
\author{
Nicholas Hamilton
}
