% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-legend-position.R
\name{theme_legend_position}
\alias{theme_legend_position}
\title{Position Legend in Convenient Locations}
\usage{
theme_legend_position(x = "topleft")
}
\arguments{
\item{x}{the position, valid values are topleft, middleleft, bottomleft, topright, middleright and bottomright, or the shortened versions 
respecitvely, tl, ml, bl, tr, mr, br}
}
\description{
A convenience function to position the legend at various internal positions
}
\author{
Nicholas Hamilton
}
