\name{theme_rgbg}
\alias{theme_rgb}
\alias{theme_rgbg}
\alias{theme_tern_rgbg}
\title{A theme with grey background, red, green and blue axes and white gridlines}
\usage{
  theme_rgbg(base_size = 12, base_family = "")
}
\description{
  \code{theme_rgbg} is a theme with grey background, red,
  green and blue axes and white gridlines
}
\seealso{
  \code{\link{theme_rgbw}}
}

