% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-rotate.R
\name{theme_rotate}
\alias{theme_rotate}
\title{Rotate Ternary Diagram}
\usage{
theme_rotate(degrees = 60, radians = degrees * pi/180)
}
\arguments{
\item{degrees, radians}{specify the angle to rotate the plot by in either degrees or radians. 
If both \code{degrees} and \code{radians} are specified, then precedence is given to 
the \code{radians} argument. If no value is specified, the plot will rotate by 60 degrees}
}
\description{
Convenience function to rotate the diagram by an angle in degrees or radians.
}
\examples{
x = ggtern(data.frame(x=1,y=1,z=1),aes(x,y,z))
for(a in seq(0,60,by=15))
print(x + theme_rotate(a))
}
\author{
Nicholas Hamilton
}
