% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-noarrows.R
\name{theme_noarrows}
\alias{theme_noarrows}
\alias{theme_hidearrows}
\alias{theme_showarrows}
\title{Show or Hide the Ternary Arrows}
\usage{
theme_noarrows()

theme_hidearrows()

theme_showarrows()
}
\description{
\code{theme_noarrows} is a function that appends to the current theme a flag to switch OFF the ternary arrows
}
\author{
Nicholas Hamilton
}
