% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-nomask.R
\name{theme_nomask}
\alias{theme_nomask}
\alias{theme_hidemask}
\alias{theme_showmask}
\title{Show or Hide the Clipping Mask}
\usage{
theme_nomask()

theme_hidemask()

theme_showmask()
}
\description{
Convenience Function to Show or Hide the Clipping Mask, \code{theme_showmask} is a function that 
appends to the current theme a flag to switch ON the clipping mask, whilst, \code{theme_nomask} (or \code{theme_hidemask}) is a function 
that appends to the current theme a flag to switch OFF the clipping mask
}
\author{
Nicholas Hamilton
}
