% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-ticks.R
\name{theme_ticklength}
\alias{theme_ticklength}
\alias{theme_ticklength_major}
\alias{theme_ticklength_minor}
\title{Modify the Ticklengths}
\usage{
theme_ticklength(major = NULL, minor = NULL)

theme_ticklength_major(major)

theme_ticklength_minor(minor)
}
\arguments{
\item{major, minor}{lenth of major and minor ticklengths respectively. 
Must be a 'rel' (ie ex rel(5.0)), or will be ignored.}
}
\description{
Convenience Function for changing the major and/or minor ticklengths.
}
\examples{
ggtern() +
  theme_ticklength(major = rel(1.0),
                   minor = rel(0.5))
}
\author{
Nicholas Hamilton
}
