% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-ticksoutside.R
\name{theme_ticksoutside}
\alias{theme_ticksoutside}
\alias{theme_ticksinside}
\title{Place Ticks Inside or Outside}
\usage{
theme_ticksoutside()

theme_ticksinside()
}
\description{
\code{theme_ticksoutside} is a function that ensures the ticks are placed OUTSIDE of the plot area, whereas, 
\code{theme_ticksinside} is a function that ensures the ticks are placed INSIDE of the plot area 
(opposite to \code{theme_ticksoutside})
}
\author{
Nicholas Hamilton
}
