% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-.R
\name{transform_position_tern}
\alias{transform_position_tern}
\title{Convenience function to transform all position variables in a ternary plot}
\usage{
transform_position_tern(
  df,
  trans_x = NULL,
  trans_y = NULL,
  trans_z = NULL,
  ...
)
}
\arguments{
\item{trans_x, trans_y, trans_z}{Transformation functions for x, y and z aesthetics.
(will transform x, xmin, xmax, xend etc)}

\item{...}{Additional arguments passed to \code{trans_x} and \code{trans_y}.}
}
\description{
Convenience function to transform all position variables in a ternary plot
}
\author{
Nicholas Hamilton
}
\keyword{internal}
