% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{undoCartesian}
\alias{undoCartesian}
\title{Undo a Ternary Transformation}
\usage{
undoCartesian(df, coord)
}
\arguments{
\item{df}{data frame containing x, y, z}

\item{coord}{ternary coordinate system}
}
\description{
Using the provided ternary coordinates, and a data-frame containing x, y, z values, convert back
to the cartesian coordinates
}

