\name{unit}
\alias{unit}
\title{\code{unit} is an exported instance of the imported grid unit function.
For further information, refer to: \link[grid]{unit}.}
\description{
  \code{unit} is an exported instance of the imported grid
  unit function. For further information, refer to:
  \link[grid]{unit}.
}

