% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depreciated.R
\name{zzz-depreciated}
\alias{zzz-depreciated}
\alias{tern_stop}
\alias{clipPolygons}
\alias{polyclip}
\alias{theme_arrowbaseline}
\alias{element_ternary}
\alias{ggtern.multi}
\alias{multi}
\alias{multiplot}
\alias{point.in.sequence}
\title{Depreciated Functions}
\source{
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\usage{
tern_stop(src = "target")

clipPolygons(
  df,
  coord,
  plyon = c("level", "piece", "group"),
  op = "intersection"
)

theme_arrowbaseline(label = "labels")

element_ternary(
  showarrows,
  padding,
  arrowsep,
  arrowstart,
  arrowfinish,
  vshift,
  hshift,
  ticklength.major,
  ticklength.minor
)

ggtern.multi(..., plotlist = NULL, cols = 1, layout = NULL)

point.in.sequence(x, y, ..., df = data.frame(x = x, y = y), close = FALSE)
}
\arguments{
\item{src}{character name of current procedure}

\item{df}{a data frame}

\item{coord}{a ternary coordinate system}

\item{plyon}{items in the data frame to pass to ddply argument}

\item{op}{operation method to clip, intersection, union, minus or xor}

\item{label}{a character ('axis','ticks' or 'labels') or numeric (rounded to 0, 1 or 2) value to determine the relative location (labels is default)
if a character is provided, and it is not one of the above, an error will be thrown.}

\item{showarrows}{logical whether to show the axis directional arrows DEPRECIATED}

\item{padding}{the padding around the plot area to make provision for axis labels, ticks and arrows, relative to the cartesian plane. DEPRECIATED}

\item{arrowsep}{the distance between ternary axis and ternary arrows DEPRECIATED}

\item{arrowstart}{the proportion along the ternary axis to start the directional arrow DEPRECIATED}

\item{arrowfinish}{the proportion along the ternary axis to stop the directional arrow DEPRECIATED}

\item{vshift}{shift the plot area vertically DEPRECIATED}

\item{hshift}{shift the plot area horizontally DEPRECIATED}

\item{ticklength.major}{the length of the major ternary ticks as an euclidean distance 
relative to the x and y limits of the cartesian plot area. DEPRECIATED}

\item{ticklength.minor}{the length of the minor ternary ticks as an euclidean distance 
relative to the x and y limits of the cartesian plot area. DEPRECIATED}

\item{...}{additional arguments, multiple plot objects}

\item{plotlist}{alternative to the ... argument, provide a list of ggplot or grob objects, objects which do not inherit 
the ggplot or grob classes will be stripped.}

\item{cols}{number of columns if the layout parameter is not provided.}

\item{layout}{override number of cols, and provide a matrix specifying the layout}

\item{x}{vector of numeric \code{x} values}

\item{y}{vector of numeric \code{y} values}

\item{close}{logical value (default \code{FALSE}), as to whether the set should be closed by adding (duplicating) 
the first row (after ordering) to the end of the set.}
}
\value{
\code{data.frame} object containing the re-ordered input set.
}
\description{
The following is a list of functions which were once used in previous versions of ggtern, however,
have now been depreciated

\strong{DEPRECIATED:} \code{tern_stop(\dots)} Internal Function, checks if the most recent coordinate system is ternary, and, if not, 
stops the current procedure, with a common message format

\strong{DEPRECIATED:} \code{clipPolygons(\dots)} Using the using the PolyClip Package, This clips input polygons for 
use in the density and contour geometries.

\strong{DEPRECIATED:} \code{theme_arrowbaseline(\dots)} The ternary arrows can have an offset unit value (see \code{tern.axis.arrow.sep}), 
however, it is convenient to set this relative to either the axis, ticks or axis ticklabels (since the latter
 two can be hidden / removed.). This function permits this to be set

\strong{DEPRECIATED:} \code{element_ternary(\dots)} Replaced by individual theme elements:
\enumerate{
  \item \code{tern.axis.arrow.show}
  \item \code{tern.axis.padding}
  \item \code{tern.axis.arrow.sep}
  \item \code{tern.axis.arrow.start}
  \item \code{tern.axis.arrow.finish}
  \item \code{tern.axis.vshift}
  \item \code{tern.axis.hshift}
  \item \code{tern.axis.ticks.length.major}
  \item \code{tern.axis.ticks.length.minor}
}

\strong{DEPRECIATED:} \code{ggtern.multi} is a function which permits the arrangement of muliple \code{ggtern} or \code{ggplot2} objects,
plots can be provided to the elipsis argument, or, as a list and at the simplest case, the number of columns can be
specified. For more advanced usage, consider the layout argument.

\strong{DEPRECIATED:} The \code{point.in.sequence} function takes numeric input vectors \code{x} and \code{y} or a 
\code{\link{data.frame}} object, and orders the values in such way that they are correctly sequenced by the angle subtended between each point,
and, the centroid of the total set. If the data is provided in the format of a \code{data.frame}, then it must 
containing columns named \code{x} and \code{y}, else an error will be thrown.
}
\details{
Used to define the layout of some of the ggtern plot features which are unique to the ternary diagrams , and hence, this package.

By default, 1 column is specified, which means that the plots will be stacked on top of each other in a single column,
however, if say 4 plots are provided to the ellipsis or \code{plotlist}, with \code{cols} equal to 2, 
then this will produce a 2 x 2 arrangement.

In regards to the \code{layout} argument (which overrides the \code{cols} argument), if it is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot number 1 will go in the upper left, 2 will go in the upper right, and 3 will go all the way across the 
bottom - see the last example below.

The arguments \code{x} and \code{y} represent cartesian coordinates. This is useful if a path is sought that 
passes through each point in the ordered set, however, no two lines in the total path cross over each other. 
Uses the \code{\link{atan2}} function to determine the angle (theta) between each point (x,y) and the centroid 
of the data, it then orders based on increasing values of theta.
}
\author{
Nicholas Hamilton
}
\keyword{clipping}
\keyword{depreciated}
\keyword{polygon}
