% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-internal.R
\name{.getFunctions}
\alias{.getFunctions}
\title{OLD FUNCTIONS  
new_panel','train_layout','train_position','train_ranges','map_position','map_layout','reset_scales','facet_render',
xlabel','ylabel'
expand_default', ## REMOVED}
\usage{
.getFunctions()
}
\description{
OLD FUNCTIONS  
new_panel','train_layout','train_position','train_ranges','map_position','map_layout','reset_scales','facet_render',
xlabel','ylabel'
expand_default', ## REMOVED
}
