% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/element_ternary.R
\name{element_ternary}
\alias{element_ternary}
\title{Theme element: ternary structure}
\usage{
element_ternary(showarrows, padding, arrowsep, arrowstart, arrowfinish, vshift,
  hshift, ticklength.major, ticklength.minor)
}
\arguments{
\item{showarrows}{logical whether to show the axis directional arrows DEPRECIATED}

\item{padding}{the padding around the plot area to make provision for axis labels, ticks and arrows, relative to the cartesian plane. DEPRECIATED}

\item{arrowsep}{the distance between ternary axis and ternary arrows DEPRECIATED}

\item{arrowstart}{the proportion along the ternary axis to start the directional arrow DEPRECIATED}

\item{arrowfinish}{the proportion along the ternary axis to stop the directional arrow DEPRECIATED}

\item{vshift}{shift the plot area vertically DEPRECIATED}

\item{hshift}{shift the plot area horizontally DEPRECIATED}

\item{ticklength.major}{the length of the major ternary ticks as an euclidean distance
relative to the x and y limits of the cartesian plot area. DEPRECIATED}

\item{ticklength.minor}{the length of the minor ternary ticks as an euclidean distance
relative to the x and y limits of the cartesian plot area. DEPRECIATED}
}
\description{
\strong{NB:} This function has been entirely \strong{DEPRECIATED} as of version 1.0.1.3 and, has been replaced by individual theme elements:
\enumerate{
  \item \code{\link{axis.tern.showarrows}}
  \item \code{\link{axis.tern.padding}}
  \item \code{\link{axis.tern.arrowsep}}
  \item \code{\link{axis.tern.arrowstart}}
  \item \code{\link{axis.tern.arrowfinish}}
  \item \code{\link{axis.tern.vshift}}
  \item \code{\link{axis.tern.hshift}}
  \item \code{\link{axis.tern.ticklength.major}}
  \item \code{\link{axis.tern.ticklength.minor}}
}
}
\details{
Used to define the layout of some of the ggtern plot features which are unique to the ternary diagrams , and hence, this package.
}
\keyword{internal}

