% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getBreaks}
\alias{getBreaks}
\alias{getLabels}
\title{Generate Axis Breaks and Labels}
\usage{
getBreaks(limits = c(0, 1), isMajor = TRUE, n = 5)

getLabels(limits = c(0, 1), breaks = getBreaks(limits), format = "\%g",
  factor = 100)
}
\arguments{
\item{limits}{the scale limits}

\item{isMajor}{major or minor grids}

\item{n}{number of breaks}

\item{breaks}{numeric denoting the breaks}

\item{format}{the formatting string to be passed through to the \code{\link{sprintf}} function}

\item{factor}{the multiplicative factor}
}
\description{
Calculates the Breaks for Major or Minor Gridlines based 
on the input limits.
}

