% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtern-constructor.R
\name{ggtern}
\alias{constructor}
\alias{ggtern}
\title{ggtern Constructor}
\usage{
ggtern(...)
}
\arguments{
\item{...}{Other arguments passed on to methods. Not currently used.}
}
\value{
\code{ggtern(...)} returns an object of class \code{ggplot}.
}
\description{
Plots in \code{ggtern} are instigated via the default constructor: \code{ggtern(...)}, which is essentially a convenience wrapper for the following: 
\code{ggplot{...} + coord_tern()}, indeed, if one wishes to use \code{ggplot{...} + coord_tern()} then this is quite satisfactory.
}
\seealso{
For an introduction to the \code{ggtern} package, (including many examples), click \link[=ggtern-package]{HERE}.
}

