% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{zzz-overloaded}
\alias{zzz-overloaded}
\title{Overloaded ggplot2 functions}
\format{
functions and objects
}
\description{
INTERNAL FUNCTIONS (Overloaded from ggplot2): The source of the following functions originate 
from ggplot2, however, minor patches were required in order for them to function under the ggtern framework. 
Patches were mainly to do with handling the new theme elements and heirarchies.
}
\keyword{internal}
