% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-jitter-tern.R
\docType{data}
\name{position_jitter_tern}
\alias{PositionJitterTern}
\alias{position_jitter_tern}
\title{Jitter ternary points to avoid overplotting.}
\usage{
position_jitter_tern(x = NULL, y = NULL, z = NULL)
}
\arguments{
\item{x, y, z}{amount of positional jitter}
}
\description{
Jitter ternary points to avoid overplotting.
}
\seealso{
Other position.adjustments: \code{\link{position_nudge_tern}}
}
\keyword{datasets}

