% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rgb2hex}
\alias{rgb2hex}
\title{Convert RGB to HEX Color}
\usage{
rgb2hex(r = 0, g = 0, b = 0)
}
\arguments{
\item{r, g, b}{colors, numeric scalar between 0 and 255}
}
\description{
Function to convert rgb color to hex color
}
\examples{
rgb2hex(0,0,0)
rgb2hex(255,255,255)
rgb2hex(255,0,0)
rgb2hex(0,255,0) 
rgb2hex(0,0,255)
}
\author{
Nicholas Hamilton
}
\keyword{internal}

