\name{theme_nocart}
\alias{theme_noarrows}
\alias{theme_nocart}
\alias{theme_showarrows}
\alias{theme_tern_bw}
\alias{theme_tern_gray}
\alias{theme_tern_rgbg}
\alias{theme_tern_rgbw}
\title{Ternary Theme Elements}
\usage{
  theme_nocart()

  theme_noarrows()

  theme_showarrows()

  theme_tern_rgbg()

  theme_tern_rgbw()

  theme_tern_bw()

  theme_tern_gray()
}
\description{
  \code{theme_nocart} is a function that returns empty
  theme elemens for the standard cartesian memebers,
  including panel.background, panel.border,
  panel.grid.major, panel.grid.minor, axis.ticks,
  axis.text.x, axis.text.y, axis.title.x, axis.title.y are
  all set to blank

  \code{theme_noarrows} is a function that apepnds to the
  current theme a flag to switch OFF the ternary arrows

  \code{theme_showarrows} is a function that apepnds to the
  current theme a flag to switch ON the ternary arrows

  \code{theme_tern_rgbg} ternary theme, red green blue with
  gray background

  \code{theme_tern_rgbw} ternary theme, red green blue with
  white background

  \code{theme_tern_bw} ternary theme black and white

  \code{theme_tern_gray} ternary theme, gray theme
}

