% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme}
\alias{theme}
\title{Set theme elements (ggtern version)}
\usage{
theme(..., complete = FALSE, validate = TRUE)
}
\arguments{
\item{...}{a list of element name, element pairings that modify the
existing theme.}

\item{complete}{set this to TRUE if this is a complete theme, such as
the one returned \code{by theme_grey()}. Complete themes behave
differently when added to a ggplot object.}

\item{validate}{TRUE to run validate_element, FALSE to bypass checks.}
}
\description{
Use this function to modify theme settings.
}
\author{
Nicholas Hamilton
}

