% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tern-clockwise.R
\name{theme_clockwise}
\alias{tern_anticlockwise}
\alias{tern_clockwise}
\alias{tern_counterclockwise}
\alias{theme_anticlockwise}
\alias{theme_clockwise}
\alias{theme_counterclockwise}
\title{Direction of Ternary Rotation}
\usage{
theme_clockwise()

theme_anticlockwise()

theme_counterclockwise()

tern_clockwise()

tern_anticlockwise()

tern_counterclockwise()
}
\description{
\code{theme_clockwise, theme_anticlockwise} (or their aliasses) are function that instructs the axes
precession to be clockwise or anticlockwise respectively.
}
\details{
If the \code{\link{axis.tern.showarrows}} value is \code{FALSE}, these functions will set it to \code{TRUE}.
}

