% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-clockwise.R
\name{theme_clockwise}
\alias{theme_clockwise}
\alias{theme_anticlockwise}
\alias{theme_counterclockwise}
\alias{tern_clockwise}
\alias{tern_anticlockwise}
\alias{tern_counterclockwise}
\title{Direction of Ternary Rotation}
\usage{
theme_clockwise()

theme_anticlockwise()

theme_counterclockwise()
}
\description{
\code{theme_clockwise, theme_anticlockwise} (or their aliasses) are function that instructs the axes 
precession to be clockwise or anticlockwise respectively.
}
\details{
If the \code{tern.axis.arrow.show} value is \code{FALSE}, these functions will set it to \code{TRUE}.
}
\author{
Nicholas Hamilton
}
