% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-defaults.R
\name{theme_complete}
\alias{theme_complete}
\title{List of Available Themes}
\description{
\code{ggtern} ships with a number of complete themes, summarized as follows. 
These themes combine the base themes available to \code{ggplot2} and a number of NEW themes, which
are unique to \code{ggtern}.
\itemize{
 \item{ Black and White Theme:}{ 
   \code{\link[=theme_bw]{theme_bw(...)}}
 }
 \item{Minimal Theme:}{
   \code{\link[=theme_minimal]{theme_minimal(...)}}
 }
 \item{Classic Theme:}{
   \code{\link[=theme_classic]{theme_classic(...)}}
 }
 \item{Gray and White Theme:}{
   \code{\link[=theme_gray]{theme_gray(...)}}
 }
 \item{Red, Green, Blue and White Theme:}{
   \code{\link[=theme_rgbw]{theme_rgbw(...)}}
 }
 \item{Red, Green, Blue and Gray Theme:}{
   \code{\link[=theme_rgbg]{theme_rgbg(...)}}
 }
 \item{Dark Theme:}{
   \code{\link[=theme_dark]{theme_dark(...)}}
 }
 \item{Darker Theme:}{
   \code{\link[=theme_darker]{theme_darker(...)}}
 }
 \item{Light Theme:}{
   \code{\link[=theme_light]{theme_light(...)}}
 }
 \item{Theme with Only Black Lines:}{
   \code{\link[=theme_linedraw]{theme_linedraw(...)}}
 }
 \item{Matrix Theme:}{
   \code{\link[=theme_matrix]{theme_matrix(...)}}
 }
 \item{Tropical Theme:}{
   \code{\link[=theme_tropical]{theme_tropical(...)}}
 }
 \item{BlueLight Theme:}{
   \code{\link[=theme_bluelight]{theme_bluelight(...)}}
 }
 \item{BlueDark Theme:}{
   \code{\link[=theme_bluedark]{theme_bluedark(...)}}
 }
 \item{Black Vermillion Blue Theme (White Background):}{
   \code{\link[=theme_bvbw]{theme_bvbw(...)}}
 }
 \item{Black Vermillion Blue Theme (Grey Background):}{
   \code{\link[=theme_bvbg]{theme_bvbg(...)}}
 }
}
}
\seealso{
\code{\link{ggtern_themes}}
}
\author{
Nicholas Hamilton
}
