% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-elements.R
\name{theme_elements}
\alias{theme_elements}
\title{New Theme Elements}
\description{
\code{ggtern} creates many new theme elements and inheritances, the following is an outline:
}
\details{
Theme elements can inherit properties from other theme elements.
For example, \code{axis.title.x} inherits from \code{axis.title}, 
which in turn inherits from \code{text}. All text elements inherit
directly or indirectly from \code{text}; all lines inherit from
\code{line}, and all rectangular objects inherit from \code{rect}.

Modifying the newly created items requires the same procedures as introduced in the ggplot2 \code{\link[ggplot2]{theme}} documentation.
Some convenience functions have been also newly created, proceed to \code{\link{theme_convenience_functions}} for additional information.
}
\section{New/Additional Inheritance Structures}{

\Sexpr[results=rd,stage=build]{ggtern:::rd_theme()}
**  \strong{NB:} \code{tern.panel.background}, whilst the ternary area is 'triangular' per-se, \code{\link{element_rect}} has been used, 
as it actually holds NO information regarding the geometry (width, height), only fill, color, 
size and linetype border (ie the style of how it will be rendered).
}

\author{
Nicholas Hamilton
}
