% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-gridsontop.R
\name{theme_gridsontop}
\alias{theme_gridsontop}
\title{Render Grids on Top}
\usage{
theme_gridsontop()
}
\description{
Convenience function to render the major and minor grids on top of the other layers.
}
\author{
Nicholas Hamilton
}

