% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-mesh.R
\name{theme_mesh}
\alias{theme_mesh}
\title{Create Grid Mesh}
\usage{
theme_mesh(number = 5)
}
\arguments{
\item{number}{the number of major breaks.}
}
\description{
Convenience function for creation of a grid mesh of 'n' major breaks.
}

