% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme-defaults.R
\name{theme_minimal}
\alias{theme_minimal}
\alias{theme_tern_minimal}
\title{A minimalistic theme with no background annotations (ggtern version).}
\usage{
theme_minimal(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
A minimalistic theme with no background annotations (ggtern version).
}

