% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme-defaults.R
\name{theme_rgbg}
\alias{theme_rgb}
\alias{theme_rgbg}
\alias{theme_tern_rgbg}
\title{A theme with grey background, red, green and blue axes and white gridlines}
\usage{
theme_rgbg(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
\code{theme_rgbg} is a theme with grey background, red, green and blue axes and white gridlines
}
\seealso{
\code{\link{theme_rgbw}}
}

