% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme-defaults.R
\name{theme_rgbw}
\alias{theme_rgbw}
\alias{theme_tern_rgbw}
\title{A theme with white background, red, green and blue axes and gidlines}
\usage{
theme_rgbw(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
\code{theme_rgbw} is a theme with white background, red, green and blue axes and gridlines
}
\seealso{
\code{\link{theme_rgbg}}
}

