% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-showlabels.R
\name{theme_showlabels}
\alias{theme_showlabels}
\alias{theme_hidelabels}
\alias{theme_nolabels}
\title{Show or Hide Axis Ticklabels}
\usage{
theme_showlabels()

theme_hidelabels()

theme_nolabels()
}
\description{
Convenience functions to enable or disable the axis ticklabels
}
\details{
\code{theme_showlabels} is a function that apends to the current theme a flag to switch ON the axis ticklabels, whilst 
\code{theme_hidelabels} or \code{theme_nolabels} (Alias) are functions that apends to the current theme a flag 
to switch OFF the axis ticklabels
}
\author{
Nicholas Hamilton
}
