% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-showtitles.R
\name{theme_showtitles}
\alias{theme_showtitles}
\alias{theme_hidetitles}
\alias{theme_notitles}
\title{Show or Hide the Axis (Apex) Titles}
\usage{
theme_showtitles()

theme_hidetitles()

theme_notitles()
}
\description{
Convenience functions to SHOW or HIDE the apex labels.
}
\examples{
 #Load data
 data(Feldspar)
 ggtern(data=Feldspar,aes(An,Ab,Or)) + geom_point() + theme_bw() + theme_hidetitles()
}
\author{
Nicholas Hamilton
}
