% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-zoom.R
\name{theme_zoom_X}
\alias{theme_zoom_L}
\alias{theme_zoom_R}
\alias{theme_zoom_T}
\alias{theme_zoom_X}
\alias{theme_zoom_center}
\title{Zoom on Plot Region}
\usage{
theme_zoom_T(x = 1, ...)

theme_zoom_L(x = 1, ...)

theme_zoom_R(x = 1, ...)

theme_zoom_center(x = 1, ...)
}
\arguments{
\item{x}{numeric scalar}
}
\description{
A series of convenience functions for the zooming in on the middle or apex regions to various degrees.
In these convenience functions, a single value of \code{x} is expected, which defines the values of the apex
limits other than the point of reference, for example, \code{theme_zoom_T} will fix the \code{T} limit 
at \code{1}, and will adjust the balancing limits according to the argument x. Equivalent are also possible for
the \code{L} and \code{R} apexes, via the \code{theme_zoom_L} and \code{theme_zoom_R} functions respectively. 
Finally, the \code{theme_zoom_center} function will adjust all three apex limits, serving, as the name suggests, 
to act as a centred zoom. The examples below are fairly self explanatory.
}
\examples{
#Default Plot
data(Feldspar)
base = ggtern(Feldspar,aes(Ab,An,Or)) +
       geom_point() + 
       geom_density_tern()
base

#Zoom on Left Region
base + theme_zoom_L(0.5)

#Zoom on Right Region
base + theme_zoom_R(0.5)

#Zoom on Top Region
base + theme_zoom_T(0.5)

#Zoom on Center Region
base + theme_zoom_center(0.5)

}
\author{
Nicholas Hamilton
}

