% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-noarrows.R, R/theme-nomask.R
\name{theme_noarrows}
\alias{theme_hidearrows}
\alias{theme_hidemask}
\alias{theme_noarrows}
\alias{theme_showarrows}
\title{Show or Hide the Ternary Arrows}
\usage{
theme_noarrows()

theme_hidearrows()

theme_showarrows()

theme_hidemask()
}
\description{
\code{theme_noarrows} is a function that appends to the current theme a flag to switch OFF the ternary arrows

\code{theme_hidearrows} is an alias for \code{theme_noarrows}

\code{theme_showarrows} is a function that appends to the current theme a flag to switch ON the ternary arrows
}

