% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{zzz-internal}
\alias{zzz-internal}
\alias{ifthenelse}
\alias{is.numericor}
\alias{find_global_tern}
\alias{tern_dep}
\alias{arrow_label_formatter}
\alias{joinCharacterSeries}
\alias{identityInv}
\alias{getFormulaVars}
\alias{scales_add_missing_tern}
\alias{layers_add_or_remove_mask}
\title{Internal Functions}
\usage{
ifthenelse(x, a, b)

is.numericor(A, B)

find_global_tern(name, env = environment(), mode = "any")

tern_dep(version, msg)

arrow_label_formatter(label, suffix = NULL, sep = "/", ...)

joinCharacterSeries(x, lastWord = "and")

identityInv(z)

getFormulaVars(x, dependent = TRUE)

scales_add_missing_tern(plot)

layers_add_or_remove_mask(plot)
}
\arguments{
\item{x}{formula object}

\item{a}{value to return if \code{x} is TRUE}

\item{b}{value to return if \code{x} is FALSE}

\item{A}{value to return if numeric}

\item{B}{numeric value to return if \code{A} is NOT numeric}

\item{name}{character name of object to search for}

\item{env}{environment to search within as first priority}

\item{mode}{the mode to search within}

\item{version}{The last version of ggplot2 where this function was good
(in other words, the last version where it was not deprecated).}

\item{msg}{The message to print.}

\item{label}{character label}

\item{suffix}{chacater suffix behind each label}

\item{sep}{the seperator between label and suffix}

\item{...}{additional arguments}

\item{dependent}{whether to return the dependent variables (TRUE) or the indpenedent variables (FALSE)}

\item{plot}{ggplot object}

\item{latex}{logical as to whether latex formats should be parsed}

\item{ggplot}{object}
}
\description{
INTERNAL FUNCTIONS: \code{ggtern} makes use of several non-exported internal functions, list are as follows:
}
\author{
Nicholas Hamilton
}
\keyword{internal}
